Two total_duration-instead-gap_size bugs and a JSON_Parser int-double bug smashed.
Yes, I've smashed em for good.

 On branch dev
 Your branch is up to date with 'origin/dev'.

 Changes to be committed:
	new file:   appointment_request_test4.json
	new file:   appointment_request_test5.json
	new file:   appointment_test4.json
	modified:   appointy/src/request_handling.cc
	modified:   appointy_base/src/util/json_parser.cc
	modified:   tests/test.cc

+++ appointy_base/src/util/json_parser.cc
@@ -367,7 +367,7 @@ auto JSON_Parser::parse_answer(const nlohmann::json &answer) -> Answer *
     {
         try
         {
-            return new NumericAnswer<int> {answer_signature_id, answer.at("number")};
+            return new NumericAnswer<double> {answer_signature_id, answer.at("number")};
         }
         catch(const nlohmann::detail::out_of_range &)
         {
